refactor: move release ID generation from client to server
This change refactors the release ID generation logic to be handled
server-side instead of client-side, improving architecture and
enabling atomic operations.

Changes:
- Remove --release-title flag (redundant with release ID template)
- Update CreateReleaseRequest proto to accept release_id_template and
  release_id_timezone instead of pre-computed train
- Move renderTrain() and validateTemplate() from action layer to
  backend/api/v1
- Client now passes template and timezone; server computes train
- Plan title now derives from release ID instead of separate title
- Move tests from action/command to backend/api/v1

Benefits:
- Server owns release ID generation logic (single source of truth)
- Train computation happens atomically with iteration increment
- Cleaner client API (configuration vs implementation)
- Better separation of concerns
