+++ frontend/src/types/proto-es/v1/release_service_pb.js
@@ -16,7 +16,7 @@ import { file_v1_sql_service } from "./sql_service_pb";
  * Describes the file v1/release_service.proto.
  */
 export const file_v1_release_service = /*@__PURE__*/
-  fileDesc("Chh2MS9yZWxlYXNlX3NlcnZpY2UucHJvdG8SC2J5dGViYXNlLnYxIj8KEUdldFJlbGVhc2VSZXF1ZXN0EioKBG5hbWUYASABKAlCHOBBAvpBFgoUYnl0ZWJhc2UuY29tL1JlbGVhc2UikAEKE0xpc3RSZWxlYXNlc1JlcXVlc3QSLAoGcGFyZW50GAEgASgJQhzgQQL6QRYKFGJ5dGViYXNlLmNvbS9Qcm9qZWN0EhEKCXBhZ2Vfc2l6ZRgCIAEoBRISCgpwYWdlX3Rva2VuGAMgASgJEhQKDHNob3dfZGVsZXRlZBgEIAEoCBIOCgZmaWx0ZXIYBSABKAkiVwoUTGlzdFJlbGVhc2VzUmVzcG9uc2USJgoIcmVsZWFzZXMYASADKAsyFC5ieXRlYmFzZS52MS5SZWxlYXNlEhcKD25leHRfcGFnZV90b2tlbhgCIAEoCSJ/ChRDcmVhdGVSZWxlYXNlUmVxdWVzdBIsCgZwYXJlbnQYASABKAlCHOBBAvpBFgoUYnl0ZWJhc2UuY29tL1Byb2plY3QSKgoHcmVsZWFzZRgCIAEoCzIULmJ5dGViYXNlLnYxLlJlbGVhc2VCA+BBAhINCgV0cmFpbhgDIAEoCSJzChRVcGRhdGVSZWxlYXNlUmVxdWVzdBIqCgdyZWxlYXNlGAEgASgLMhQuYnl0ZWJhc2UudjEuUmVsZWFzZUID4EECEi8KC3VwZGF0ZV9tYXNrGAIgASgLMhouZ29vZ2xlLnByb3RvYnVmLkZpZWxkTWFzayJCChREZWxldGVSZWxlYXNlUmVxdWVzdBIqCgRuYW1lGAEgASgJQhzgQQL6QRYKFGJ5dGViYXNlLmNvbS9SZWxlYXNlIkQKFlVuZGVsZXRlUmVsZWFzZVJlcXVlc3QSKgoEbmFtZRgBIAEoCUIc4EEC+kEWChRieXRlYmFzZS5jb20vUmVsZWFzZSKWAQoTQ2hlY2tSZWxlYXNlUmVxdWVzdBIsCgZwYXJlbnQYASABKAlCHOBBAvpBFgoUYnl0ZWJhc2UuY29tL1Byb2plY3QSKgoHcmVsZWFzZRgCIAEoCzIULmJ5dGViYXNlLnYxLlJlbGVhc2VCA+BBAhIPCgd0YXJnZXRzGAMgAygJEhQKDGN1c3RvbV9ydWxlcxgEIAEoCSKwAgoUQ2hlY2tSZWxlYXNlUmVzcG9uc2USPgoHcmVzdWx0cxgBIAMoCzItLmJ5dGViYXNlLnYxLkNoZWNrUmVsZWFzZVJlc3BvbnNlLkNoZWNrUmVzdWx0EhUKDWFmZmVjdGVkX3Jvd3MYAiABKAMSKgoKcmlza19sZXZlbBgDIAEoDjIWLmJ5dGViYXNlLnYxLlJpc2tMZXZlbBqUAQoLQ2hlY2tSZXN1bHQSDAoEZmlsZRgBIAEoCRIOCgZ0YXJnZXQYAiABKAkSJAoHYWR2aWNlcxgDIAMoCzITLmJ5dGViYXNlLnYxLkFkdmljZRIVCg1hZmZlY3RlZF9yb3dzGAQgASgDEioKCnJpc2tfbGV2ZWwYBSABKA4yFi5ieXRlYmFzZS52MS5SaXNrTGV2ZWwiTAocTGlzdFJlbGVhc2VDYXRlZ29yaWVzUmVxdWVzdBIsCgZwYXJlbnQYASABKAlCHOBBAvpBFgoUYnl0ZWJhc2UuY29tL1Byb2plY3QiMwodTGlzdFJlbGVhc2VDYXRlZ29yaWVzUmVzcG9uc2USEgoKY2F0ZWdvcmllcxgBIAMoCSKEBQoHUmVsZWFzZRIRCgRuYW1lGAEgASgJQgPgQQMSEAoIY2F0ZWdvcnkYAiABKAkSKAoFZmlsZXMYAyADKAsyGS5ieXRlYmFzZS52MS5SZWxlYXNlLkZpbGUSMgoKdmNzX3NvdXJjZRgEIAEoCzIeLmJ5dGViYXNlLnYxLlJlbGVhc2UuVkNTU291cmNlEhQKB2NyZWF0b3IYBSABKAlCA+BBAxI0CgtjcmVhdGVfdGltZRgGIAEoCzIaLmdvb2dsZS5wcm90b2J1Zi5UaW1lc3RhbXBCA+BBAxImCgVzdGF0ZRgHIAEoDjISLmJ5dGViYXNlLnYxLlN0YXRlQgPgQQMSJwoEdHlwZRgIIAEoDjIZLmJ5dGViYXNlLnYxLlJlbGVhc2UuVHlwZRqWAQoERmlsZRIMCgRwYXRoGAIgASgJEg8KB3ZlcnNpb24YBiABKAkSFAoMZW5hYmxlX2dob3N0GAkgASgIEiYKBXNoZWV0GAMgASgJQhf6QRQKEmJ5dGViYXNlLmNvbS9TaGVldBIWCglzdGF0ZW1lbnQYByABKAxCA+BBBBIZCgxzaGVldF9zaGEyNTYYBCABKAlCA+BBAxpACglWQ1NTb3VyY2USJgoIdmNzX3R5cGUYASABKA4yFC5ieXRlYmFzZS52MS5WQ1NUeXBlEgsKA3VybBgCIAEoCSI8CgRUeXBlEhQKEFRZUEVfVU5TUEVDSUZJRUQQABINCglWRVJTSU9ORUQQARIPCgtERUNMQVJBVElWRRACOkDqQT0KFGJ5dGViYXNlLmNvbS9SZWxlYXNlEiVwcm9qZWN0cy97cHJvamVjdH0vcmVsZWFzZXMve3JlbGVhc2V9MsAKCg5SZWxlYXNlU2VydmljZRKKAQoKR2V0UmVsZWFzZRIeLmJ5dGViYXNlLnYxLkdldFJlbGVhc2VSZXF1ZXN0GhQuYnl0ZWJhc2UudjEuUmVsZWFzZSJG2kEEbmFtZYrqMA9iYi5yZWxlYXNlcy5nZXSQ6jABgtPkkwIiEiAvdjEve25hbWU9cHJvamVjdHMvKi9yZWxlYXNlcy8qfRKeAQoMTGlzdFJlbGVhc2VzEiAuYnl0ZWJhc2UudjEuTGlzdFJlbGVhc2VzUmVxdWVzdBohLmJ5dGViYXNlLnYxLkxpc3RSZWxlYXNlc1Jlc3BvbnNlIknaQQZwYXJlbnSK6jAQYmIucmVsZWFzZXMubGlzdJDqMAGC0+STAiISIC92MS97cGFyZW50PXByb2plY3RzLyp9L3JlbGVhc2VzEqYBCg1DcmVhdGVSZWxlYXNlEiEuYnl0ZWJhc2UudjEuQ3JlYXRlUmVsZWFzZVJlcXVlc3QaFC5ieXRlYmFzZS52MS5SZWxlYXNlIlzaQQ5wYXJlbnQscmVsZWFzZYrqMBJiYi5yZWxlYXNlcy5jcmVhdGWQ6jABgtPkkwIrOgdyZWxlYXNlIiAvdjEve3BhcmVudD1wcm9qZWN0cy8qfS9yZWxlYXNlcxKzAQoNVXBkYXRlUmVsZWFzZRIhLmJ5dGViYXNlLnYxLlVwZGF0ZVJlbGVhc2VSZXF1ZXN0GhQuYnl0ZWJhc2UudjEuUmVsZWFzZSJp2kETcmVsZWFzZSx1cGRhdGVfbWFza4rqMBJiYi5yZWxlYXNlcy51cGRhdGWQ6jABgtPkkwIzOgdyZWxlYXNlMigvdjEve3JlbGVhc2UubmFtZT1wcm9qZWN0cy8qL3JlbGVhc2VzLyp9EpUBCg1EZWxldGVSZWxlYXNlEiEuYnl0ZWJhc2UudjEuRGVsZXRlUmVsZWFzZVJlcXVlc3QaFi5nb29nbGUucHJvdG9idWYuRW1wdHkiSdpBBG5hbWWK6jASYmIucmVsZWFzZXMuZGVsZXRlkOowAYLT5JMCIiogL3YxL3tuYW1lPXByb2plY3RzLyovcmVsZWFzZXMvKn0SmwEKD1VuZGVsZXRlUmVsZWFzZRIjLmJ5dGViYXNlLnYxLlVuZGVsZXRlUmVsZWFzZVJlcXVlc3QaFC5ieXRlYmFzZS52MS5SZWxlYXNlIk2K6jAUYmIucmVsZWFzZXMudW5kZWxldGWQ6jABgtPkkwIrIikvdjEve25hbWU9cHJvamVjdHMvKi9yZWxlYXNlcy8qfTp1bmRlbGV0ZRKfAQoMQ2hlY2tSZWxlYXNlEiAuYnl0ZWJhc2UudjEuQ2hlY2tSZWxlYXNlUmVxdWVzdBohLmJ5dGViYXNlLnYxLkNoZWNrUmVsZWFzZVJlc3BvbnNlIkqK6jARYmIucmVsZWFzZXMuY2hlY2uQ6jABgtPkkwIrOgEqIiYvdjEve3BhcmVudD1wcm9qZWN0cy8qfS9yZWxlYXNlczpjaGVjaxLIAQoVTGlzdFJlbGVhc2VDYXRlZ29yaWVzEikuYnl0ZWJhc2UudjEuTGlzdFJlbGVhc2VDYXRlZ29yaWVzUmVxdWVzdBoqLmJ5dGViYXNlLnYxLkxpc3RSZWxlYXNlQ2F0ZWdvcmllc1Jlc3BvbnNlIljaQQZwYXJlbnSK6jAQYmIucmVsZWFzZXMubGlzdJDqMAGC0+STAjESLy92MS97cGFyZW50PXByb2plY3RzLyp9L3JlbGVhc2VzOmxpc3RDYXRlZ29yaWVzQqkBCg9jb20uYnl0ZWJhc2UudjFCE1JlbGVhc2VTZXJ2aWNlUHJvdG9QAVo0Z2l0aHViLmNvbS9ieXRlYmFzZS9ieXRlYmFzZS9iYWNrZW5kL2dlbmVyYXRlZC1nby92MaICA0JYWKoCC0J5dGViYXNlLlYxygILQnl0ZWJhc2VcVjHiAhdCeXRlYmFzZVxWMVxHUEJNZXRhZGF0YeoCDEJ5dGViYXNlOjpWMWIGcHJvdG8z", [file_google_api_annotations, file_google_api_client, file_google_api_field_behavior, file_google_api_resource, file_google_protobuf_empty, file_google_protobuf_field_mask, file_google_protobuf_timestamp, file_v1_annotation, file_v1_common, file_v1_sql_service]);
+  fileDesc("Chh2MS9yZWxlYXNlX3NlcnZpY2UucHJvdG8SC2J5dGViYXNlLnYxIj8KEUdldFJlbGVhc2VSZXF1ZXN0EioKBG5hbWUYASABKAlCHOBBAvpBFgoUYnl0ZWJhc2UuY29tL1JlbGVhc2UikAEKE0xpc3RSZWxlYXNlc1JlcXVlc3QSLAoGcGFyZW50GAEgASgJQhzgQQL6QRYKFGJ5dGViYXNlLmNvbS9Qcm9qZWN0EhEKCXBhZ2Vfc2l6ZRgCIAEoBRISCgpwYWdlX3Rva2VuGAMgASgJEhQKDHNob3dfZGVsZXRlZBgEIAEoCBIOCgZmaWx0ZXIYBSABKAkiVwoUTGlzdFJlbGVhc2VzUmVzcG9uc2USJgoIcmVsZWFzZXMYASADKAsyFC5ieXRlYmFzZS52MS5SZWxlYXNlEhcKD25leHRfcGFnZV90b2tlbhgCIAEoCSKqAQoUQ3JlYXRlUmVsZWFzZVJlcXVlc3QSLAoGcGFyZW50GAEgASgJQhzgQQL6QRYKFGJ5dGViYXNlLmNvbS9Qcm9qZWN0EioKB3JlbGVhc2UYAiABKAsyFC5ieXRlYmFzZS52MS5SZWxlYXNlQgPgQQISGwoTcmVsZWFzZV9pZF90ZW1wbGF0ZRgDIAEoCRIbChNyZWxlYXNlX2lkX3RpbWV6b25lGAQgASgJInMKFFVwZGF0ZVJlbGVhc2VSZXF1ZXN0EioKB3JlbGVhc2UYASABKAsyFC5ieXRlYmFzZS52MS5SZWxlYXNlQgPgQQISLwoLdXBkYXRlX21hc2sYAiABKAsyGi5nb29nbGUucHJvdG9idWYuRmllbGRNYXNrIkIKFERlbGV0ZVJlbGVhc2VSZXF1ZXN0EioKBG5hbWUYASABKAlCHOBBAvpBFgoUYnl0ZWJhc2UuY29tL1JlbGVhc2UiRAoWVW5kZWxldGVSZWxlYXNlUmVxdWVzdBIqCgRuYW1lGAEgASgJQhzgQQL6QRYKFGJ5dGViYXNlLmNvbS9SZWxlYXNlIpYBChNDaGVja1JlbGVhc2VSZXF1ZXN0EiwKBnBhcmVudBgBIAEoCUIc4EEC+kEWChRieXRlYmFzZS5jb20vUHJvamVjdBIqCgdyZWxlYXNlGAIgASgLMhQuYnl0ZWJhc2UudjEuUmVsZWFzZUID4EECEg8KB3RhcmdldHMYAyADKAkSFAoMY3VzdG9tX3J1bGVzGAQgASgJIrACChRDaGVja1JlbGVhc2VSZXNwb25zZRI+CgdyZXN1bHRzGAEgAygLMi0uYnl0ZWJhc2UudjEuQ2hlY2tSZWxlYXNlUmVzcG9uc2UuQ2hlY2tSZXN1bHQSFQoNYWZmZWN0ZWRfcm93cxgCIAEoAxIqCgpyaXNrX2xldmVsGAMgASgOMhYuYnl0ZWJhc2UudjEuUmlza0xldmVsGpQBCgtDaGVja1Jlc3VsdBIMCgRmaWxlGAEgASgJEg4KBnRhcmdldBgCIAEoCRIkCgdhZHZpY2VzGAMgAygLMhMuYnl0ZWJhc2UudjEuQWR2aWNlEhUKDWFmZmVjdGVkX3Jvd3MYBCABKAMSKgoKcmlza19sZXZlbBgFIAEoDjIWLmJ5dGViYXNlLnYxLlJpc2tMZXZlbCJMChxMaXN0UmVsZWFzZUNhdGVnb3JpZXNSZXF1ZXN0EiwKBnBhcmVudBgBIAEoCUIc4EEC+kEWChRieXRlYmFzZS5jb20vUHJvamVjdCIzCh1MaXN0UmVsZWFzZUNhdGVnb3JpZXNSZXNwb25zZRISCgpjYXRlZ29yaWVzGAEgAygJIoQFCgdSZWxlYXNlEhEKBG5hbWUYASABKAlCA+BBAxIQCghjYXRlZ29yeRgCIAEoCRIoCgVmaWxlcxgDIAMoCzIZLmJ5dGViYXNlLnYxLlJlbGVhc2UuRmlsZRIyCgp2Y3Nfc291cmNlGAQgASgLMh4uYnl0ZWJhc2UudjEuUmVsZWFzZS5WQ1NTb3VyY2USFAoHY3JlYXRvchgFIAEoCUID4EEDEjQKC2NyZWF0ZV90aW1lGAYgASgLMhouZ29vZ2xlLnByb3RvYnVmLlRpbWVzdGFtcEID4EEDEiYKBXN0YXRlGAcgASgOMhIuYnl0ZWJhc2UudjEuU3RhdGVCA+BBAxInCgR0eXBlGAggASgOMhkuYnl0ZWJhc2UudjEuUmVsZWFzZS5UeXBlGpYBCgRGaWxlEgwKBHBhdGgYAiABKAkSDwoHdmVyc2lvbhgGIAEoCRIUCgxlbmFibGVfZ2hvc3QYCSABKAgSJgoFc2hlZXQYAyABKAlCF/pBFAoSYnl0ZWJhc2UuY29tL1NoZWV0EhYKCXN0YXRlbWVudBgHIAEoDEID4EEEEhkKDHNoZWV0X3NoYTI1NhgEIAEoCUID4EEDGkAKCVZDU1NvdXJjZRImCgh2Y3NfdHlwZRgBIAEoDjIULmJ5dGViYXNlLnYxLlZDU1R5cGUSCwoDdXJsGAIgASgJIjwKBFR5cGUSFAoQVFlQRV9VTlNQRUNJRklFRBAAEg0KCVZFUlNJT05FRBABEg8KC0RFQ0xBUkFUSVZFEAI6QOpBPQoUYnl0ZWJhc2UuY29tL1JlbGVhc2USJXByb2plY3RzL3twcm9qZWN0fS9yZWxlYXNlcy97cmVsZWFzZX0ywAoKDlJlbGVhc2VTZXJ2aWNlEooBCgpHZXRSZWxlYXNlEh4uYnl0ZWJhc2UudjEuR2V0UmVsZWFzZVJlcXVlc3QaFC5ieXRlYmFzZS52MS5SZWxlYXNlIkbaQQRuYW1liuowD2JiLnJlbGVhc2VzLmdldJDqMAGC0+STAiISIC92MS97bmFtZT1wcm9qZWN0cy8qL3JlbGVhc2VzLyp9Ep4BCgxMaXN0UmVsZWFzZXMSIC5ieXRlYmFzZS52MS5MaXN0UmVsZWFzZXNSZXF1ZXN0GiEuYnl0ZWJhc2UudjEuTGlzdFJlbGVhc2VzUmVzcG9uc2UiSdpBBnBhcmVudIrqMBBiYi5yZWxlYXNlcy5saXN0kOowAYLT5JMCIhIgL3YxL3twYXJlbnQ9cHJvamVjdHMvKn0vcmVsZWFzZXMSpgEKDUNyZWF0ZVJlbGVhc2USIS5ieXRlYmFzZS52MS5DcmVhdGVSZWxlYXNlUmVxdWVzdBoULmJ5dGViYXNlLnYxLlJlbGVhc2UiXNpBDnBhcmVudCxyZWxlYXNliuowEmJiLnJlbGVhc2VzLmNyZWF0ZZDqMAGC0+STAis6B3JlbGVhc2UiIC92MS97cGFyZW50PXByb2plY3RzLyp9L3JlbGVhc2VzErMBCg1VcGRhdGVSZWxlYXNlEiEuYnl0ZWJhc2UudjEuVXBkYXRlUmVsZWFzZVJlcXVlc3QaFC5ieXRlYmFzZS52MS5SZWxlYXNlImnaQRNyZWxlYXNlLHVwZGF0ZV9tYXNriuowEmJiLnJlbGVhc2VzLnVwZGF0ZZDqMAGC0+STAjM6B3JlbGVhc2UyKC92MS97cmVsZWFzZS5uYW1lPXByb2plY3RzLyovcmVsZWFzZXMvKn0SlQEKDURlbGV0ZVJlbGVhc2USIS5ieXRlYmFzZS52MS5EZWxldGVSZWxlYXNlUmVxdWVzdBoWLmdvb2dsZS5wcm90b2J1Zi5FbXB0eSJJ2kEEbmFtZYrqMBJiYi5yZWxlYXNlcy5kZWxldGWQ6jABgtPkkwIiKiAvdjEve25hbWU9cHJvamVjdHMvKi9yZWxlYXNlcy8qfRKbAQoPVW5kZWxldGVSZWxlYXNlEiMuYnl0ZWJhc2UudjEuVW5kZWxldGVSZWxlYXNlUmVxdWVzdBoULmJ5dGViYXNlLnYxLlJlbGVhc2UiTYrqMBRiYi5yZWxlYXNlcy51bmRlbGV0ZZDqMAGC0+STAisiKS92MS97bmFtZT1wcm9qZWN0cy8qL3JlbGVhc2VzLyp9OnVuZGVsZXRlEp8BCgxDaGVja1JlbGVhc2USIC5ieXRlYmFzZS52MS5DaGVja1JlbGVhc2VSZXF1ZXN0GiEuYnl0ZWJhc2UudjEuQ2hlY2tSZWxlYXNlUmVzcG9uc2UiSorqMBFiYi5yZWxlYXNlcy5jaGVja5DqMAGC0+STAis6ASoiJi92MS97cGFyZW50PXByb2plY3RzLyp9L3JlbGVhc2VzOmNoZWNrEsgBChVMaXN0UmVsZWFzZUNhdGVnb3JpZXMSKS5ieXRlYmFzZS52MS5MaXN0UmVsZWFzZUNhdGVnb3JpZXNSZXF1ZXN0GiouYnl0ZWJhc2UudjEuTGlzdFJlbGVhc2VDYXRlZ29yaWVzUmVzcG9uc2UiWNpBBnBhcmVudIrqMBBiYi5yZWxlYXNlcy5saXN0kOowAYLT5JMCMRIvL3YxL3twYXJlbnQ9cHJvamVjdHMvKn0vcmVsZWFzZXM6bGlzdENhdGVnb3JpZXNCqQEKD2NvbS5ieXRlYmFzZS52MUITUmVsZWFzZVNlcnZpY2VQcm90b1ABWjRnaXRodWIuY29tL2J5dGViYXNlL2J5dGViYXNlL2JhY2tlbmQvZ2VuZXJhdGVkLWdvL3YxogIDQlhYqgILQnl0ZWJhc2UuVjHKAgtCeXRlYmFzZVxWMeICF0J5dGViYXNlXFYxXEdQQk1ldGFkYXRh6gIMQnl0ZWJhc2U6OlYxYgZwcm90bzM", [file_google_api_annotations, file_google_api_client, file_google_api_field_behavior, file_google_api_resource, file_google_protobuf_empty, file_google_protobuf_field_mask, file_google_protobuf_timestamp, file_v1_annotation, file_v1_common, file_v1_sql_service]);
 
 /**
  * Describes the message bytebase.v1.GetReleaseRequest.
