+++ proto/gen/grpc-doc/v1/index.html
@@ -23651,11 +23651,22 @@ Should be a database. Format: instances/{instance}/databases/{database} </p></td
                 </tr>
               
                 <tr>
-                  <td>train</td>
+                  <td>release_id_template</td>
                   <td><a href="#string">string</a></td>
                   <td></td>
-                  <td><p>Train for iteration tracking (template rendered without {iteration}).
-Used to group releases and determine the next iteration number. </p></td>
+                  <td><p>Template for release ID generation.
+Available variables: {date}, {time}, {timestamp}, {iteration}.
+Example: &#34;release_{date}-RC{iteration}&#34; generates &#34;release_20260119-RC00&#34;.
+Default: &#34;release_{date}-RC{iteration}&#34;. </p></td>
+                </tr>
+              
+                <tr>
+                  <td>release_id_timezone</td>
+                  <td><a href="#string">string</a></td>
+                  <td></td>
+                  <td><p>Timezone for {date} and {time} variables in the template.
+Must be a valid IANA timezone (e.g., &#34;UTC&#34;, &#34;America/Los_Angeles&#34;).
+Default: &#34;UTC&#34;. </p></td>
                 </tr>
               
             </tbody>
